+++ dali/c_api/c_api.cc
@@ -62,19 +62,19 @@ int PopCurrBatchSize(batch_size_map_t *batch_size_map, int max_batch_size,
 }
 
 /**
- * @brief Extract ExtSrcNoCopyMode based on the DALI_ext_force_copy and DALI_ext_force_no_copy
+ * @brief Extract InputOperatorSettingMode based on the DALI_ext_force_copy and DALI_ext_force_no_copy
  *
  * @param flags Flags typically specified in daliSetExternalInput* functions.
  */
-dali::ExtSrcNoCopyMode GetExternalSourceCopyMode(unsigned int flags) {
-  dali::ExtSrcNoCopyMode no_copy_mode = dali::ExtSrcNoCopyMode::DEFAULT;
+dali::InputOperatorNoCopyMode GetExternalSourceCopyMode(unsigned int flags) {
+  dali::InputOperatorNoCopyMode no_copy_mode = dali::InputOperatorNoCopyMode::DEFAULT;
   DALI_ENFORCE(!((flags & DALI_ext_force_copy) && (flags & DALI_ext_force_no_copy)),
                "External Source cannot be forced to use DALI_ext_force_copy and "
                "DALI_ext_force_no_copy at the same time.");
   if (flags & DALI_ext_force_copy) {
-    no_copy_mode = dali::ExtSrcNoCopyMode::FORCE_COPY;
+    no_copy_mode = dali::InputOperatorNoCopyMode::FORCE_COPY;
   } else if (flags & DALI_ext_force_no_copy) {
-    no_copy_mode = dali::ExtSrcNoCopyMode::FORCE_NO_COPY;
+    no_copy_mode = dali::InputOperatorNoCopyMode::FORCE_NO_COPY;
   }
   return no_copy_mode;
 }
