+++ dali/pipeline/operator/builtin/external_source.cc
@@ -20,42 +20,22 @@ namespace dali {
 
 template <>
 void ExternalSource<CPUBackend>::RunImpl(Workspace &ws) {
-  std::list<uptr_tl_type> tensor_list_elm;
-  {
-    std::unique_lock<std::mutex> busy_lock(busy_m_);
-    tensor_list_elm = tl_data_.PopFront();
-    state_.pop_front();
-  }
   auto &output = ws.Output<CPUBackend>(0);
-  // if the output is pinned and input not it needs to be copied
-  if (output.is_pinned() && !tensor_list_elm.front()->is_pinned()) {
-    auto &thread_pool = ws.GetThreadPool();
-    const auto &shapes = tensor_list_elm.front()->shape();
-    auto curr_batch_size = shapes.num_samples();
-    output.Resize(shapes, tensor_list_elm.front()->type());
+  auto &thread_pool = ws.GetThreadPool();
+  ForwardCurrentData(output, thread_pool);
+}
 
-    // as we copy element by element and the output is contiguous we need to set layout
-    // for the whole output not each element(view)
-    auto &output = ws.Output<CPUBackend>(0);
-    output.SetLayout(tensor_list_elm.front()->GetLayout());
 
-    for (int sample_id = 0; sample_id < curr_batch_size; ++sample_id) {
-      thread_pool.AddWork(
-          [&output, sample_id, &tensor_list_elm](int tid) {
-            output.CopySample(sample_id, *tensor_list_elm.front(), sample_id, AccessOrder::host());
-          },
-          shapes.tensor_size(sample_id));
-    }
-    thread_pool.RunAll();
-  } else {
-    // swap output with tensor_list_elm content
-    std::swap(output, *tensor_list_elm.front());
-  }
-  RecycleBuffer(tensor_list_elm);
+template<>
+void ExternalSource<GPUBackend>::RunImpl(Workspace &ws) {
+  auto &output = ws.Output<GPUBackend>(0);
+  cudaStream_t stream_used = ws.has_stream() ? ws.stream() : 0;
+  ForwardCurrentData(output, stream_used);
 }
 
 
 DALI_REGISTER_OPERATOR(ExternalSource, ExternalSource<CPUBackend>, CPU);
+DALI_REGISTER_OPERATOR(ExternalSource, ExternalSource<GPUBackend>, GPU);
 
 
 // This schema is partially internal. We want it to be listed int the supported_ops,
@@ -72,26 +52,7 @@ DALI_SCHEMA(ExternalSource)
   or by the Pipeline::AddExternalInput method.)code")
   .NumInput(0)
   .NumOutput(1)
-  .AddOptionalArg("blocking",
-      R"code(Whether external source should block until data is available or just
-fail when it is not)code", true)
-  .AddOptionalArg("no_copy",
-      R"code(Determines whether DALI should copy the buffer when feed_input is called.
-
-If set to True, DALI passes the user's memory directly to the pipeline, instead of copying it.
-It is the user's responsibility to keep the buffer alive and unmodified until it is
-consumed by the pipeline.
-
-The buffer can be modified or freed again after the outputs of the relevant iterations
-have been consumed. Effectively, it happens after ``prefetch_queue_depth`` or
-``cpu_queue_depth * gpu_queue_depth`` (when they are not equal) iterations following
-the``feed_input`` call.
-
-The memory location must match the specified ``device`` parameter of the operator.
-For the CPU, the provided memory can be one contiguous buffer or a list of contiguous Tensors.
-For the GPU, to avoid extra copy, the provided buffer must be contiguous. If you provide a list
-of separate Tensors, there will be an additional copy made internally, consuming both memory
-and bandwidth.)code", false)
+  .AddParent("InputOperatorBase")
   .AddOptionalTypeArg("dtype", R"code(Input data type.
 
 The operator will validate that the fetched data is of the provided type.
